# docs/swagger.yaml
basePath: /
definitions:
  dto.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        minLength: 8
        type: string
      new_password_confirmation:
        type: string
    required:
    - current_password
    - new_password
    - new_password_confirmation
    type: object
  dto.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      request_id:
        type: string
    type: object
  dto.ForgotPasswordRequest:
    properties:
      email:
        type: string
      tenant_id:
        type: string
    required:
    - email
    - tenant_id
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      tenant_id:
        type: string
    required:
    - email
    - password
    - tenant_id
    type: object
  dto.LogoutRequest:
    properties:
      refresh_token:
        type: string
    type: object
  dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
  dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
      tenant_id:
        type: string
    required:
    - refresh_token
    - tenant_id
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      password_confirmation:
        type: string
      tenant_id:
        type: string
    required:
    - email
    - password
    - password_confirmation
    - tenant_id
    type: object
  dto.RegisterResponse:
    properties:
      credential_id:
        type: string
      message:
        type: string
    type: object
  dto.ResetPasswordRequest:
    properties:
      email:
        type: string
      new_password:
        minLength: 8
        type: string
      new_password_confirmation:
        type: string
      token:
        type: string
    required:
    - email
    - new_password
    - new_password_confirmation
    - token
    type: object
  dto.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Authentication and authorization microservice for TravApp platform.
    Provides user registration, login, OAuth2 (Google), JWT token management, password
    management, and email verification.
  title: TravApp Auth Service API
  version: "1.0"
paths:
  /api/v1/auth/.well-known/jwks.json:
    get:
      description: Returns the public keys used to verify JWT tokens in JWKS format
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get JSON Web Key Set
      tags:
      - Token
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password, returns access and refresh
        tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: User login
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidates the current access token and optionally the refresh
        token
      parameters:
      - description: Logout request with optional refresh token
        in: body
        name: request
        schema:
          $ref: '#/definitions/dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout current session
      tags:
      - Auth
  /api/v1/auth/logout/all:
    post:
      description: Invalidates all active sessions and tokens for the authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout all sessions
      tags:
      - Auth
  /api/v1/auth/oauth/google:
    get:
      description: Redirects the user to Google's OAuth2 consent page for authentication
      parameters:
      - description: Tenant ID (UUID format)
        in: query
        name: tenant_id
        required: true
        type: string
      responses:
        "302":
          description: Redirect to Google OAuth2
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Initiate Google OAuth2 login
      tags:
      - OAuth
  /api/v1/auth/oauth/google/callback:
    get:
      description: Handles the callback from Google OAuth2, exchanges code for tokens
      parameters:
      - description: Authorization code from Google
        in: query
        name: code
        required: true
        type: string
      - description: OAuth2 state parameter
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Google OAuth2 callback
      tags:
      - OAuth
  /api/v1/auth/password/change:
    post:
      consumes:
      - application/json
      description: Change the password for the currently authenticated user
      parameters:
      - description: Change password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Password
  /api/v1/auth/password/forgot:
    post:
      consumes:
      - application/json
      description: Sends a password reset link to the user's email address
      parameters:
      - description: Forgot password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Request password reset
      tags:
      - Password
  /api/v1/auth/password/reset:
    post:
      consumes:
      - application/json
      description: Reset user's password using a valid password reset token
      parameters:
      - description: Reset password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Reset password
      tags:
      - Password
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account with email and password
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: User registration
      tags:
      - Auth
  /api/v1/auth/token/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a new pair of access and refresh
        tokens
      parameters:
      - description: Refresh token request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Refresh access token
      tags:
      - Token
  /api/v1/auth/verify-email:
    get:
      description: Verify user's email address using the verification token sent via
        email
      parameters:
      - description: Email verification token
        in: query
        name: token
        required: true
        type: string
      - description: User's email address
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Verify email address
      tags:
      - Password
  /health:
    get:
      description: Check the health status of the service, database, and Redis
      produces:
      - application/json
      responses:
        "200":
          description: Service is healthy
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service is unhealthy
          schema:
            additionalProperties: true
            type: object
      summary: Health check
      tags:
      - Health
securityDefinitions:
  BearerAuth:
    description: Paste access_token saja (tanpa prefix Bearer), otomatis ditambahkan.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"